requirements-gpu.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile --output-file=requirements-gpu.txt requirements-gpu.in
#

absl-py==1.1.0
    # via
    #   tensorboard
    #   tensorflow-estimator
    #   tensorflow-gpu
astor==0.8.1
    # via tensorflow-gpu
backcall==0.2.0
    # via ipython
cached-property==1.5.2
    # via h5py
cycler==0.11.0
    # via matplotlib
dataclasses==0.8
    # via werkzeug
decorator==4.4.2
    # via
    #   ipython
    #   networkx
    #   traitlets
gast==0.5.3
    # via tensorflow-gpu
grpcio==1.46.3
    # via
    #   tensorboard
    #   tensorflow-gpu
h5py==3.1.0
    # via keras-applications
imageio==2.15.0
    # via scikit-image
importlib-metadata==4.8.3
    # via markdown
ipython==7.16.3
    # via -r requirements-gpu.in
ipython-genutils==0.2.0
    # via traitlets
jedi==0.17.2
    # via ipython
keras-applications==1.0.8
    # via tensorflow-gpu
keras-preprocessing==1.1.2
    # via tensorflow-gpu
kiwisolver==1.3.1
    # via matplotlib
markdown==3.3.7
    # via tensorboard
matplotlib==3.3.4
    # via scikit-image
mock==4.0.3
    # via tensorflow-estimator
networkx==2.5.1
    # via scikit-image
numpy==1.19.5
    # via
    #   -r requirements-gpu.in
    #   h5py
    #   imageio
    #   keras-applications
    #   keras-preprocessing
    #   matplotlib
    #   opencv-python
    #   pywavelets
    #   scipy
    #   tensorboard
    #   tensorflow-estimator
    #   tensorflow-gpu
opencv-python==4.5.5.64
    # via -r requirements-gpu.in
parso==0.7.1
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.4.0
    # via
    #   -r requirements-gpu.in
    #   imageio
    #   matplotlib
    #   scikit-image
prompt-toolkit==3.0.29
    # via ipython
protobuf==3.19.4
    # via
    #   tensorboard
    #   tensorflow-gpu
ptyprocess==0.7.0
    # via pexpect
pygments==2.12.0
    # via ipython
pyparsing==3.0.7
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pywavelets==1.1.1
    # via scikit-image
scikit-image==0.16.2
    # via -r requirements-gpu.in
scipy==1.5.4
    # via scikit-image
six==1.16.0
    # via
    #   grpcio
    #   keras-preprocessing
    #   python-dateutil
    #   tensorboard
    #   tensorflow-estimator
    #   tensorflow-gpu
    #   traitlets
tensorboard==1.13.1
    # via tensorflow-gpu
tensorflow-estimator==1.13.0
    # via tensorflow-gpu
tensorflow-gpu==1.13.1
    # via -r requirements-gpu.in
termcolor==1.1.0
    # via tensorflow-gpu
traitlets==4.3.3
    # via ipython
typing-extensions==4.1.1
    # via importlib-metadata
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.3
    # via tensorboard
wheel==0.37.1
    # via
    #   tensorboard
    #   tensorflow-gpu
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools